Create e-Manifest Development Q&A-20190625 1800-1.mp4
Recording of the 6/25/19 Developers call.
Topics discussed
Brokers and Transporters wire frames and review of 6/11 call.
6/26 Preproduction Release Contents:
1)	Swagger has been upgraded to the latest version: https://rcrainfopreprod.epa.gov/rcrainfo/secured/swagger/
a.	States testing services, you can still use the e-Manifest services labeled Industry, we are working on relabeling the services for State specific usage as well.
2)	Get services
a.	If a site is listed on a manifest as a transporter only they can retrieve manifests via the GET Manifest service.
3)	Save/Update/Correct manifest services:
a.	Transporter IDs are set to a maximum of 12 alphanumeric characters
b.	Prevent Ship dates older than 6/30/2018
c.	Improved error messaging for locations which were throwing a phone number error "Mandatory Field is Not Provided","field": "Emanifest.generator.contact.phone.number” when the site should have been listed as not available to e-Manifest.  In order to upload these sites, you will need to provide the full handler information for those sites just like any other site not in e-Manifest.
4)	Provided an export to CSV a list of manifests listed on an Invoice
5)	Improvements to the clarity of the invoice page and PDF based on user feedback, more clearly identifying changes due to adjustments
6)	Front end changes include:
a.	Generator Name has been added to the All Manifest view on the dashboard,
b.	The above the changes from the Save/Update/Correct manifest on the front end