Added Can eSign and Broker to the schema
Can eSign:  Indicates if there is at least one user registered for this site which can sign electronically
And adding Broker to the Site Type for a site.  Please note, this is only if the site is only a broker, if they are also a generator, transporter, or TSDF then that type will show.